python_requirements.txt
colorama
keyboard
gitpython
requests
winotify
pygame
pysocks
aiohttp
pydub
pyaudio
numpy
pywin32
typing_extensions
pyautogui
Pillow
opencv-python
pyglet
pymem
pynput
pywifi
comtypes
phonenumbers
opencage
pycountry
geopy
pycaw
tkcalendar
currency_converter
forex-python
googletrans
click
pygments
rich
chess
neat-python
wmi
SpeechRecognition
pyttsx3
speedtest-cli
